Refactored file_fixes and tests into separate modules
