+++ pre_proc/file_fix/__init__.py
@@ -1,0 +1,8 @@
+"""
+file_fix/__init__.py
+
+The workers that fix the netCDF files
+"""
+from .attribute_update import *
+from .copy_attribute import *
+from .attribute_add import *
