+++ pre_proc/file_fix/attribute_update.py
@@ -1,0 +1,253 @@
+"""
+attribute_update.py
+
+Workers that fix the netCDF files that are based on the AttributeUpdate
+abstract base class.
+"""
+import os
+
+from netCDF4 import Dataset
+
+from pre_proc.common import to_float, to_int
+from pre_proc.exceptions import (AttributeNotFoundError,
+                                 AttributeConversionError,
+                                 ExistingAttributeError)
+from .abstract import AttributeUpdate
+
+
+class ParentBranchTimeDoubleFix(AttributeUpdate):
+    """
+    Make the global attribute `branch_time_in_parent` a double
+    """
+    def __init__(self, filename, directory):
+        """
+        Initialise the class
+
+        :param str filename: The basename of the file to process.
+        :param str directory: The directory that the file is currently in.
+        """
+        super(ParentBranchTimeDoubleFix, self).__init__(filename, directory)
+        self.attribute_name = 'branch_time_in_parent'
+        self.attribute_visibility = 'global'
+        self.attribute_type = 'd'
+
+    def _calculate_new_value(self):
+        """
+        The new value is the existing string converted to a double.
+        """
+        try:
+            self.new_value = to_float(self.existing_value)
+        except ValueError:
+            raise AttributeConversionError(self.filename, self.attribute_name,
+                                           'float')
+
+
+class ChildBranchTimeDoubleFix(AttributeUpdate):
+    """
+    Make the global attribute `branch_time_in_child` a double
+    """
+    def __init__(self, filename, directory):
+        """
+        Initialise the class
+
+        :param str filename: The basename of the file to process.
+        :param str directory: The directory that the file is currently in.
+        """
+        super(ChildBranchTimeDoubleFix, self).__init__(filename, directory)
+        self.attribute_name = 'branch_time_in_child'
+        self.attribute_visibility = 'global'
+        self.attribute_type = 'd'
+
+    def _calculate_new_value(self):
+        """
+        The new value is the existing string converted to a double.
+        """
+        try:
+            self.new_value = to_float(self.existing_value)
+        except ValueError:
+            raise AttributeConversionError(self.filename, self.attribute_name,
+                                           'float')
+
+
+class ForcingIndexIntFix(AttributeUpdate):
+    """
+    Make the global attribute `forcing_index` an int.
+    """
+    def __init__(self, filename, directory):
+        """
+        Initialise the class
+
+        :param str filename: The basename of the file to process.
+        :param str directory: The directory that the file is currently in.
+        """
+        super(ForcingIndexIntFix, self).__init__(filename, directory)
+        self.attribute_name = 'forcing_index'
+        self.attribute_visibility = 'global'
+        self.attribute_type = 's'
+
+    def _calculate_new_value(self):
+        """
+        The new value is the existing string converted to a double.
+        """
+        try:
+            self.new_value = to_int(self.existing_value)
+        except ValueError:
+            raise AttributeConversionError(self.filename, self.attribute_name,
+                                           'int')
+
+
+class InitializationIndexIntFix(AttributeUpdate):
+    """
+    Make the global attribute `initialization_index` an int.
+    """
+    def __init__(self, filename, directory):
+        """
+        Initialise the class
+
+        :param str filename: The basename of the file to process.
+        :param str directory: The directory that the file is currently in.
+        """
+        super(InitializationIndexIntFix, self).__init__(filename, directory)
+        self.attribute_name = 'initialization_index'
+        self.attribute_visibility = 'global'
+        self.attribute_type = 's'
+
+    def _calculate_new_value(self):
+        """
+        The new value is the existing string converted to a double.
+        """
+        try:
+            self.new_value = to_int(self.existing_value)
+        except ValueError:
+            raise AttributeConversionError(self.filename, self.attribute_name,
+                                           'int')
+
+
+class PhysicsIndexIntFix(AttributeUpdate):
+    """
+    Make the global attribute `physics_index` an int.
+    """
+    def __init__(self, filename, directory):
+        """
+        Initialise the class
+
+        :param str filename: The basename of the file to process.
+        :param str directory: The directory that the file is currently in.
+        """
+        super(PhysicsIndexIntFix, self).__init__(filename, directory)
+        self.attribute_name = 'physics_index'
+        self.attribute_visibility = 'global'
+        self.attribute_type = 's'
+
+    def _calculate_new_value(self):
+        """
+        The new value is the existing string converted to a double.
+        """
+        try:
+            self.new_value = to_int(self.existing_value)
+        except ValueError:
+            raise AttributeConversionError(self.filename, self.attribute_name,
+                                           'int')
+
+
+class RealizationIndexIntFix(AttributeUpdate):
+    """
+    Make the global attribute `realization_index` an int.
+    """
+    def __init__(self, filename, directory):
+        """
+        Initialise the class
+
+        :param str filename: The basename of the file to process.
+        :param str directory: The directory that the file is currently in.
+        """
+        super(RealizationIndexIntFix, self).__init__(filename, directory)
+        self.attribute_name = 'realization_index'
+        self.attribute_visibility = 'global'
+        self.attribute_type = 's'
+
+    def _calculate_new_value(self):
+        """
+        The new value is the existing string converted to a double.
+        """
+        try:
+            self.new_value = to_int(self.existing_value)
+        except ValueError:
+            raise AttributeConversionError(self.filename, self.attribute_name,
+                                           'int')
+
+
+class FurtherInfoUrlToHttps(AttributeUpdate):
+    """
+    Change the protocol in the further_info_url attribute from HTTP to
+    HTTPS.
+    """
+    def __init__(self, filename, directory):
+        """
+        Initialise the class
+
+        :param str filename: The basename of the file to process.
+        :param str directory: The directory that the file is currently in.
+        """
+        super(FurtherInfoUrlToHttps, self).__init__(filename, directory)
+        self.attribute_name = 'further_info_url'
+        self.attribute_visibility = 'global'
+        self.attribute_type = 'c'
+
+    def _calculate_new_value(self):
+        """
+        The new value is the existing string converted to a double.
+        """
+        if not self.existing_value.startswith('http:'):
+            raise ExistingAttributeError(self.filename, self.attribute_name,
+                                         'Existing further_info_url attribute '
+                                         'does not start with http:')
+
+        self.new_value = self.existing_value.replace('http:', 'https:', 1)
+
+
+class FurtherInfoUrlAWISourceIdAndHttps(AttributeUpdate):
+    """
+    Change the protocol in the further_info_url attribute from HTTP to
+    HTTPS.
+    """
+    def __init__(self, filename, directory):
+        """
+        Initialise the class
+
+        :param str filename: The basename of the file to process.
+        :param str directory: The directory that the file is currently in.
+        """
+        super(FurtherInfoUrlAWISourceIdAndHttps, self).__init__(filename,
+                                                                directory)
+        self.attribute_name = 'further_info_url'
+        self.source_id = None
+        self.attribute_visibility = 'global'
+        self.attribute_type = 'c'
+
+    def _calculate_new_value(self):
+        """
+        The new value is the existing string converted to a double.
+        """
+        if not self.existing_value.startswith('http:'):
+            raise ExistingAttributeError(self.filename, self.attribute_name,
+                                         'Existing further_info_url attribute '
+                                         'does not start with http:')
+
+        self.new_value = self.existing_value.replace('http:', 'https:', 1)
+        self.new_value = self.new_value.replace('AWI-CM-1-0', self.source_id)
+
+    def _get_existing_value(self):
+        """
+        Get the value of the existing attribute from the current file
+        """
+        filepath = os.path.join(self.directory, self.filename)
+        with Dataset(filepath) as rootgrp:
+            self.existing_value = getattr(rootgrp, self.attribute_name, None)
+            self.source_id = getattr(rootgrp, 'source_id', None)
+
+        if self.existing_value is None:
+            raise AttributeNotFoundError(self.filename, self.attribute_name)
+
+        if self.source_id is None:
+            raise AttributeNotFoundError(self.filename, 'source_id')
