+++ pre_proc/tests/test_file_fix/test_attribute_update.py
@@ -1,7 +1,7 @@
 """
-test_file_fix.py
+test_attribute_add.py
 
-Unit tests for all FileFix concrete classes from file_fix.py
+Unit tests for all FileFix concrete classes from attribute_add.py
 """
 import subprocess
 import unittest
@@ -14,26 +14,12 @@ from pre_proc.exceptions import (AttributeNotFoundError,
                                  InstanceVariableNotDefinedError)
 from pre_proc.file_fix import (ParentBranchTimeDoubleFix,
                                ChildBranchTimeDoubleFix,
-                               FurtherInfoUrlToHttps,
-                               ParentSourceIdFromSourceId,
-                               ParentBranchTimeAdd,
-                               ChildBranchTimeAdd,
-                               BranchMethodAdd,
-                               DataSpecsVersionAdd,
-                               CellMeasuresAreacellaAdd,
-                               CellMethodsAreaTimeMeanAdd,
-                               InitializationIndexIntFix,
                                ForcingIndexIntFix,
+                               InitializationIndexIntFix,
                                PhysicsIndexIntFix,
                                RealizationIndexIntFix,
-                               SeaWaterSalinityStandardNameAdd,
-                               CellMethodsSeaAreaTimeMeanAdd,
-                               CellMeasuresAreacelloVolcelloAdd,
-                               VarUnitsToThousandths,
-                               FurtherInfoUrlAWISourceIdAndHttps,
-                               SeaSurfaceTemperatureNameAdd,
-                               FillValueFromMissingValue,
-                               CellMeasuresAreacelloAdd)
+                               FurtherInfoUrlToHttps,
+                               FurtherInfoUrlAWISourceIdAndHttps)
 
 
 class BaseTest(unittest.TestCase):
@@ -52,9 +38,16 @@ class BaseTest(unittest.TestCase):
             def __enter__(self):
                 return self
 
-        patch = mock.patch('pre_proc.file_fix.Dataset')
+        self.dataset = MockedNamespace()
+
+        patch = mock.patch('pre_proc.file_fix.abstract.Dataset')
         self.mock_dataset = patch.start()
-        self.mock_dataset.return_value = MockedNamespace()
+        self.mock_dataset.return_value = self.dataset
+        self.addCleanup(patch.stop)
+
+        patch = mock.patch('pre_proc.file_fix.attribute_update.Dataset')
+        self.mock_local_dataset = patch.start()
+        self.mock_local_dataset.return_value = self.dataset
         self.addCleanup(patch.stop)
 
 
@@ -66,7 +59,7 @@ class TestParentBranchTimeDoubleFix(BaseTest):
         exception_text = ('Cannot find attribute branch_time_in_parent in '
                           'file 1.nc')
         self.assertRaisesRegex(AttributeNotFoundError, exception_text,
-                                fix.apply_fix)
+                               fix.apply_fix)
 
     def test_wrong_attribute_type_raises(self):
         """ Test if the required attribute can't be converted """
@@ -75,7 +68,7 @@ class TestParentBranchTimeDoubleFix(BaseTest):
         exception_text = ('Cannot convert attribute branch_time_in_parent to '
                           'new type float in file 1.nc')
         self.assertRaisesRegex(AttributeConversionError, exception_text,
-                                fix.apply_fix)
+                               fix.apply_fix)
 
     def test_no_attribute_name_raises(self):
         """ Test whether missing attribute_name handled """
@@ -87,7 +80,7 @@ class TestParentBranchTimeDoubleFix(BaseTest):
         fix._calculate_new_value()
         fix.attribute_name = None
         self.assertRaisesRegex(InstanceVariableNotDefinedError,
-                                exception_text, fix._run_ncatted)
+                               exception_text, fix._run_ncatted)
 
     def test_no_visibility_raises(self):
         """ Test whether missing visibility handled """
@@ -97,7 +90,7 @@ class TestParentBranchTimeDoubleFix(BaseTest):
         exception_text = ('ParentBranchTimeDoubleFix: attribute '
                           'attribute_visibility is not defined')
         self.assertRaisesRegex(InstanceVariableNotDefinedError,
-                                exception_text, fix.apply_fix)
+                               exception_text, fix.apply_fix)
 
     def test_no_new_value_raises(self):
         """ Test whether missing new_value handled """
@@ -109,7 +102,7 @@ class TestParentBranchTimeDoubleFix(BaseTest):
         fix._calculate_new_value()
         fix.new_value = None
         self.assertRaisesRegex(InstanceVariableNotDefinedError,
-                                exception_text, fix._run_ncatted)
+                               exception_text, fix._run_ncatted)
 
     def test_subprocess_called_correctly(self):
         """
@@ -150,7 +143,7 @@ class TestChildBranchTimeDoubleFix(BaseTest):
         exception_text = ('Cannot find attribute branch_time_in_child in '
                           'file 1.nc')
         self.assertRaisesRegex(AttributeNotFoundError, exception_text,
-                                fix.apply_fix)
+                               fix.apply_fix)
 
     def test_wrong_attribute_type_raises(self):
         """ Test if the required attribute can't be converted """
@@ -159,7 +152,7 @@ class TestChildBranchTimeDoubleFix(BaseTest):
         exception_text = ('Cannot convert attribute branch_time_in_child to '
                           'new type float in file 1.nc')
         self.assertRaisesRegex(AttributeConversionError, exception_text,
-                                fix.apply_fix)
+                               fix.apply_fix)
 
     def test_subprocess_called_correctly(self):
         """
@@ -185,7 +178,7 @@ class TestInitializationIndexIntFix(BaseTest):
         exception_text = ('Cannot find attribute initialization_index in '
                           'file 1.nc')
         self.assertRaisesRegex(AttributeNotFoundError, exception_text,
-                                fix.apply_fix)
+                               fix.apply_fix)
 
     def test_wrong_attribute_type_raises(self):
         """ Test if the required attribute can't be converted """
@@ -194,7 +187,7 @@ class TestInitializationIndexIntFix(BaseTest):
         exception_text = ('Cannot convert attribute initialization_index to '
                           'new type int in file 1.nc')
         self.assertRaisesRegex(AttributeConversionError, exception_text,
-                                fix.apply_fix)
+                               fix.apply_fix)
 
     def test_subprocess_called_correctly(self):
         """
@@ -220,7 +213,7 @@ class TestForcingIndexIntFix(BaseTest):
         exception_text = ('Cannot find attribute forcing_index in '
                           'file 1.nc')
         self.assertRaisesRegex(AttributeNotFoundError, exception_text,
-                                fix.apply_fix)
+                               fix.apply_fix)
 
     def test_wrong_attribute_type_raises(self):
         """ Test if the required attribute can't be converted """
@@ -229,7 +222,7 @@ class TestForcingIndexIntFix(BaseTest):
         exception_text = ('Cannot convert attribute forcing_index to '
                           'new type int in file 1.nc')
         self.assertRaisesRegex(AttributeConversionError, exception_text,
-                                fix.apply_fix)
+                               fix.apply_fix)
 
     def test_subprocess_called_correctly(self):
         """
@@ -255,7 +248,7 @@ class TestPhysicsIndexIntFix(BaseTest):
         exception_text = ('Cannot find attribute physics_index in '
                           'file 1.nc')
         self.assertRaisesRegex(AttributeNotFoundError, exception_text,
-                                fix.apply_fix)
+                               fix.apply_fix)
 
     def test_wrong_attribute_type_raises(self):
         """ Test if the required attribute can't be converted """
@@ -264,7 +257,7 @@ class TestPhysicsIndexIntFix(BaseTest):
         exception_text = ('Cannot convert attribute physics_index to '
                           'new type int in file 1.nc')
         self.assertRaisesRegex(AttributeConversionError, exception_text,
-                                fix.apply_fix)
+                               fix.apply_fix)
 
     def test_subprocess_called_correctly(self):
         """
@@ -290,7 +283,7 @@ class TestRealizationIndexIntFix(BaseTest):
         exception_text = ('Cannot find attribute realization_index in '
                           'file 1.nc')
         self.assertRaisesRegex(AttributeNotFoundError, exception_text,
-                                fix.apply_fix)
+                               fix.apply_fix)
 
     def test_wrong_attribute_type_raises(self):
         """ Test if the required attribute can't be converted """
@@ -299,7 +292,7 @@ class TestRealizationIndexIntFix(BaseTest):
         exception_text = ('Cannot convert attribute realization_index to '
                           'new type int in file 1.nc')
         self.assertRaisesRegex(AttributeConversionError, exception_text,
-                                fix.apply_fix)
+                               fix.apply_fix)
 
     def test_subprocess_called_correctly(self):
         """
@@ -325,7 +318,7 @@ class TestFurtherInfoUrlToHttps(BaseTest):
         exception_text = ('Cannot find attribute further_info_url in '
                           'file 1.nc')
         self.assertRaisesRegex(AttributeNotFoundError, exception_text,
-                                fix.apply_fix)
+                               fix.apply_fix)
 
     def test_not_http(self):
         """ Test exception is raised if it doesn't start with http """
@@ -334,7 +327,7 @@ class TestFurtherInfoUrlToHttps(BaseTest):
         exception_text = ('Existing further_info_url attribute does not start '
                           'with http:')
         self.assertRaisesRegex(ExistingAttributeError, exception_text,
-                                fix.apply_fix)
+                               fix.apply_fix)
 
     def test_subprocess_called_correctly_further_info_url(self):
         """
@@ -364,7 +357,7 @@ class TestFurtherInfoUrlAWISourceIdAndHttps(BaseTest):
         exception_text = ('Cannot find attribute further_info_url in file '
                           '1.nc')
         self.assertRaisesRegex(AttributeNotFoundError,
-                                exception_text, fix._get_existing_value)
+                               exception_text, fix._get_existing_value)
 
     def test_no_source_id_raises(self):
         """ Test whether missing source_id handled """
@@ -376,7 +369,7 @@ class TestFurtherInfoUrlAWISourceIdAndHttps(BaseTest):
         exception_text = ('Cannot find attribute source_id in file '
                           '1.nc')
         self.assertRaisesRegex(AttributeNotFoundError,
-                                exception_text, fix._get_existing_value)
+                               exception_text, fix._get_existing_value)
 
     def test_wrong_existing_value_raises(self):
         """ Test whether invalid existing_value handled """
@@ -388,7 +381,7 @@ class TestFurtherInfoUrlAWISourceIdAndHttps(BaseTest):
                           'does not start with http:')
         fix._get_existing_value()
         self.assertRaisesRegex(ExistingAttributeError,
-                                exception_text, fix._calculate_new_value)
+                               exception_text, fix._calculate_new_value)
 
     def test_subprocess_called_correctly(self):
         """
@@ -406,265 +399,6 @@ class TestFurtherInfoUrlAWISourceIdAndHttps(BaseTest):
             "ncatted -h -a further_info_url,global,o,c,"
             "'https://furtherinfo.es-doc.org/CMIP6.AWI.AWI-CM-1-0-LR."
             "hist-1950.none.r1i1p1f002' /a/1.nc",
-            stderr=subprocess.STDOUT,
-            shell=True
-        )
-
-
-class TestParentSourceIdFromSourceId(BaseTest):
-    """ Test ParentSourceIdFromSourceId """
-    def test_no_attribute_raises(self):
-        """ Test if the required attribute isn't found in the netCDF """
-        fix = ParentSourceIdFromSourceId('1.nc', '/a')
-        exception_text = ('Cannot find attribute source_id in '
-                          'file 1.nc')
-        self.assertRaisesRegex(AttributeNotFoundError, exception_text,
-                                fix.apply_fix)
-
-    def test_subprocess_called_correctly(self):
-        """
-        Test that an external call's been made correctly for
-        ParentSourceIdFromSourceId
-        """
-        self.mock_dataset.return_value.source_id = 'some-model'
-        self.mock_dataset.return_value.parent_source_id = 'a-model'
-        fix = ParentSourceIdFromSourceId('1.nc', '/a')
-        fix.apply_fix()
-        self.mock_subprocess.assert_called_once_with(
-            "ncatted -h -a parent_source_id,global,o,c,'some-model' /a/1.nc",
-            stderr=subprocess.STDOUT,
-            shell=True
-        )
-
-
-class TestFillValueFromMissingValue(BaseTest):
-    """ Test FillValueFromMissingValue """
-    def test_no_attribute_raises(self):
-        """ Test if the required attribute isn't found in the netCDF """
-        fix = FillValueFromMissingValue('tos_blah.nc', '/a')
-        exception_text = ('Cannot find attribute tos.missing_value in '
-                          'file tos_blah.nc')
-        self.assertRaisesRegex(AttributeNotFoundError, exception_text,
-                                fix.apply_fix)
-
-    def test_subprocess_called_correctly(self):
-        """
-        Test that an external call's been made correctly for
-        FillValueFromMissingValue
-        """
-        class MissingValue(object):
-            missing_value = 1e-7
-        self.mock_dataset.return_value.variables = {'tos': MissingValue()}
-        fix = FillValueFromMissingValue('tos_gubbins.nc', '/a')
-        fix.apply_fix()
-        self.mock_subprocess.assert_called_once_with(
-            "ncatted -h -a _FillValue,tos,o,f,1e-07 /a/tos_gubbins.nc",
-            stderr=subprocess.STDOUT,
-            shell=True
-        )
-
-
-class TestParentBranchTimeAdd(BaseTest):
-    """ Test ParentBranchTimeAdd """
-    def test_subprocess_called_correctly(self):
-        """
-        Test that an external call's been made correctly for
-        ParentBranchTimeAdd
-        """
-        fix = ParentBranchTimeAdd('1.nc', '/a')
-        fix.apply_fix()
-        self.mock_subprocess.assert_called_once_with(
-            'ncatted -h -a branch_time_in_parent,global,o,d,0.0 '
-            '/a/1.nc',
-            stderr=subprocess.STDOUT,
-            shell=True
-        )
-
-
-class TestChildBranchTimeAdd(BaseTest):
-    """ Test ChildBranchTimeAdd """
-    def test_subprocess_called_correctly(self):
-        """
-        Test that an external call's been made correctly for
-        ChildBranchTimeAdd
-        """
-        fix = ChildBranchTimeAdd('1.nc', '/a')
-        fix.apply_fix()
-        self.mock_subprocess.assert_called_once_with(
-            'ncatted -h -a branch_time_in_child,global,o,d,0.0 '
-            '/a/1.nc',
-            stderr=subprocess.STDOUT,
-            shell=True
-        )
-
-
-class TestBranchMethodAdd(BaseTest):
-    """ Test BranchMethodAdd """
-    def test_subprocess_called_correctly(self):
-        """
-        Test that an external call's been made correctly for
-        ChildBranchTimeAdd
-        """
-        fix = BranchMethodAdd('1.nc', '/a')
-        fix.apply_fix()
-        self.mock_subprocess.assert_called_once_with(
-            "ncatted -h -a branch_method,global,o,c,'no parent' "
-            "/a/1.nc",
-            stderr=subprocess.STDOUT,
-            shell=True
-        )
-
-
-class TestDataSpecsVersionAdd(BaseTest):
-    """ Test DataSpecsVersion """
-    def test_subprocess_called_correctly(self):
-        """
-        Test that an external call's been made correctly for
-        ChildBranchTimeAdd
-        """
-        fix = DataSpecsVersionAdd('1.nc', '/a')
-        fix.apply_fix()
-        self.mock_subprocess.assert_called_once_with(
-            "ncatted -h -a data_specs_version,global,o,c,'01.00.23' "
-            "/a/1.nc",
-            stderr=subprocess.STDOUT,
-            shell=True
-        )
-
-
-class TestCellMeasuresAreacellaAdd(BaseTest):
-    """ Test CellMeasuresAreacellaAdd """
-    def test_subprocess_called_correctly(self):
-        """
-        Test that an external call's been made correctly for
-        CellMeasuresAreacellaAdd
-        """
-        fix = CellMeasuresAreacellaAdd('tas_components.nc', '/a')
-        fix.apply_fix()
-        self.mock_subprocess.assert_called_once_with(
-            "ncatted -h -a cell_measures,tas,o,c,'area: areacella' "
-            "/a/tas_components.nc",
-            stderr=subprocess.STDOUT,
-            shell=True
-        )
-
-
-class TestCellMeasuresAreacelloAdd(BaseTest):
-    """ Test CellMeasuresAreacelloAdd """
-    def test_subprocess_called_correctly(self):
-        """
-        Test that an external call's been made correctly for
-        CellMeasuresAreacelloAdd
-        """
-        fix = CellMeasuresAreacelloAdd('tos_components.nc', '/a')
-        fix.apply_fix()
-        self.mock_subprocess.assert_called_once_with(
-            "ncatted -h -a cell_measures,tos,o,c,'area: areacello' "
-            "/a/tos_components.nc",
-            stderr=subprocess.STDOUT,
-            shell=True
-        )
-
-
-class TestCellMeasuresAreacelloVolcelloAdd(BaseTest):
-    """ Test CellMeasuresAreacelloVolcelloAdd """
-    def test_subprocess_called_correctly(self):
-        """
-        Test that an external call's been made correctly for
-        CellMeasuresAreacelloVolcelloAdd
-        """
-        fix = CellMeasuresAreacelloVolcelloAdd('so_components.nc', '/a')
-        fix.apply_fix()
-        self.mock_subprocess.assert_called_once_with(
-            "ncatted -h -a cell_measures,so,o,c,"
-            "'area: areacello volume: volcello' "
-            "/a/so_components.nc",
-            stderr=subprocess.STDOUT,
-            shell=True
-        )
-
-
-class TestCellMethodsAreaTimeMeanAdd(BaseTest):
-    """ Test CellMethodsAreaTimeMeanAdd """
-    def test_subprocess_called_correctly(self):
-        """
-        Test that an external call's been made correctly for
-        CellMethodsAreaTimeMeanAdd
-        """
-        fix = CellMethodsAreaTimeMeanAdd('tas_components.nc', '/a')
-        fix.apply_fix()
-        self.mock_subprocess.assert_called_once_with(
-            "ncatted -h -a cell_methods,tas,o,c,'area: time: mean' "
-            "/a/tas_components.nc",
-            stderr=subprocess.STDOUT,
-            shell=True
-        )
-
-
-class TestCellMethodsSeaAreaTimeMeanAdd(BaseTest):
-    """ Test CellMethodsSeaAreaTimeMeanAdd """
-    def test_subprocess_called_correctly(self):
-        """
-        Test that an external call's been made correctly for
-        CellMethodsSeaAreaTimeMeanAdd
-        """
-        fix = CellMethodsSeaAreaTimeMeanAdd('so_components.nc', '/a')
-        fix.apply_fix()
-        self.mock_subprocess.assert_called_once_with(
-            "ncatted -h -a cell_methods,so,o,c,"
-            "'area: mean where sea time: mean' "
-            "/a/so_components.nc",
-            stderr=subprocess.STDOUT,
-            shell=True
-        )
-
-
-class TestSeaWaterSalinityStandardNameAdd(BaseTest):
-    """ Test SeaWaterSalinityStandardNameAdd """
-    def test_subprocess_called_correctly(self):
-        """
-        Test that an external call's been made correctly for
-        SeaWaterSalinityStandardNameAdd
-        """
-        fix = SeaWaterSalinityStandardNameAdd('so_components.nc', '/a')
-        fix.apply_fix()
-        self.mock_subprocess.assert_called_once_with(
-            "ncatted -h -a standard_name,so,o,c,'sea_water_salinity' "
-            "/a/so_components.nc",
-            stderr=subprocess.STDOUT,
-            shell=True
-        )
-
-
-class TestSeaSurfaceTemperatureNameAdd(BaseTest):
-    """ Test SeaSurfaceTemperatureNameAdd """
-    def test_subprocess_called_correctly(self):
-        """
-        Test that an external call's been made correctly for
-        SeaSurfaceTemperatureNameAdd
-        """
-        fix = SeaSurfaceTemperatureNameAdd('tos_components.nc', '/a')
-        fix.apply_fix()
-        self.mock_subprocess.assert_called_once_with(
-            "ncatted -h -a standard_name,tos,o,c,'sea_surface_temperature' "
-            "/a/tos_components.nc",
-            stderr=subprocess.STDOUT,
-            shell=True
-        )
-
-
-class TestVarUnitsToThousandths(BaseTest):
-    """ Test VarUnitsToThousandths """
-    def test_subprocess_called_correctly(self):
-        """
-        Test that an external call's been made correctly for
-        VarUnitsToThousandths.
-        """
-        fix = VarUnitsToThousandths('so_components.nc', '/a')
-        fix.apply_fix()
-        self.mock_subprocess.assert_called_once_with(
-            "ncatted -h -a units,so,o,c,'0.001' "
-            "/a/so_components.nc",
             stderr=subprocess.STDOUT,
             shell=True
         )
